+++ R/datatables.R
@@ -71,7 +71,12 @@
 #'   \code{"cell"}) to enable editing a single cell. Alternatively, you can set
 #'   it to \code{"row"} to be able to edit a row, or \code{"column"} to edit a
 #'   column, or \code{"all"} to edit all cells on the current page of the table.
-#'   In all modes, start editing by doubleclicking on a cell.
+#'   In all modes, start editing by doubleclicking on a cell. This argument can
+#'   also be a list of the form \code{list(target = TARGET, disable =
+#'   list(columns = INDICES))}, where \code{TARGET} can be \code{cell},
+#'   \code{row}, \code{column}, or \code{all}, and \code{INDICES} is an integer
+#'   vector of column indices. Use the list form if you want to disable editing
+#'   certain columns.
 #' @note You are recommended to escape the table content for security reasons
 #'   (e.g. XSS attacks) when using this function in Shiny or any other dynamic
 #'   web applications.
@@ -208,7 +213,8 @@ datatable = function(
   params$caption = captionString(caption)
 
   if (isTRUE(editable)) editable = 'cell'
-  if (is.character(editable)) params$editable = editable
+  if (is.character(editable)) editable = list(target = editable, disable = list(columns = NULL))
+  if (is.list(editable)) params$editable = editable
 
   if (!identical(class(callback), class(JS(''))))
     stop("The 'callback' argument only accept a value returned from JS()")